Version 0.1.6, respect log_level, add features, more
This version fixes it so all tasks respect `:log_level`.
It also adds a methodology to set environment variables for the
containers without overridding the built-ins.
It removes the brakeman reminder.
It makes the chown/chmod file_permissions task faster.
It makes `force:restart` also run `jobs:force:restart` if jobs are
enabled.